# documents/Open-API-Specification.yaml
openapi: 3.0.0
info:
  title: BookMania User microservice - OpenAPI
  version: 1.1.0
  description: API for managing Users, Books, Reports, Connections and Statistics.
servers:
  - url: 'http://localhost:8081/c'
...
paths:
...
  /users/{username}:
    get:
      tags:
        - Search
      summary: Search for a user based on username
      description: Search for a user based on username.  This endpoint requires authentication.
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          description: The username of an AppUser
          schema:
            type: string
            example: ExampleUsername
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserModel'
        '400':
          description: Bad Request if the required parameters are not present.
        '401':
          description: Unauthorized if the JWT token is invalid
        '404':
          description: Not Found
        '500':
          description: Internal Server Error if an unexpected error occurs.

    delete:
      tags:
        - Users
      summary: An admin can delete another user's account
      description: Delete a user account by username. This endpoint requires authentication.
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          description: The username of the user to be deleted
          schema:
            type: string
            example: ExampleUsername
      responses:
        '200':
          description: OK if it was successfully deleted.
        '400':
          description: Bad Request if the required parameters are not present.
        '401':
          description: Unauthorized if the JWT token is invalid
        '404':
          description: Not Found
        '500':
          description: Internal Server Error if an unexpected error occurs.
  /users/favBook:
    get:
      tags:
        - Search
      summary: Search for users based on favorite book
      description: Search for users based on their favourite book. This endpoint requires authentication.
      security:
        - bearerAuth: []
      parameters:
        - name: favouriteBookId
          in: query
          description: The ID of a book
          schema:
            type: string
            format: uuid
            example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserModel'
        '400':
          description: Bad Request if the required parameters are not present.
        '401':
          description: Unauthorized if the JWT token is invalid
        '500':
          description: Internal Server Error if an unexpected error occurs.
  /users/favGenres:
    post:
      tags:
        - Search
      summary: Search for users based on favorite genres
      description: Search for a user based on a list of favourite genres. This endpoint requires authentication.
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Genre'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserModel'
        '400':
          description: Bad Request if the required parameters are not present.
        '401':
          description: Unauthorized if the JWT token is invalid
        '500':
          description: Internal Server Error if an unexpected error occurs.
...
components:
  schemas:
...
    UserModel:
      type: object
      properties:
        username:
          type: string
          description: The username
          example: johndoe
        email:
          type: string
          format: email
          description: The email address of the user
          example: johndoe@example.com
        name:
          type: string
          description: The full name of the user
          example: John Doe
        bio:
          type: string
          description: A brief biographical description of the user
          example: I'm an avid reader and enjoy exploring different genres of literature.
        location:
          type: string
          description: The user's location
          example: New York, NY
        favouriteGenres:
          type: array
          items:
            $ref: "#/components/schemas/Genre"
          description: The user's favorite genres of literature
          example: [
            BIOGRAPHY,
            HISTORY,
            FANTASY
          ]
        favouriteBook:
          $ref: '#/components/schemas/Book'
...
    Book:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: The unique identifier of the book
          example: "b123e456-7890-1234-5678-9abcdef01234"
        title:
          type: string
          description: The title of the book
          example: "The Adventures of Tom Sawyer"
        authors:
          type: array
          items:
            type: string
          description: An array of author names
          example: [
            "Mark Twain",
            "J.K. Rowling",
            "Harper Lee"
          ]
        genre:
          type: array
          items:
            $ref: "#/components/schemas/Genre"
          description: An array of genres
          example: [
            SCIENCE,
            MYSTERY
          ]
        description:
          type: string
          description: A brief description of the book
          example: "A classic coming-of-age story about the adventures of Tom Sawyer in a small town along the Mississippi River."
        numPages:
          type: integer
          description: The number of pages in the book
          example: 275
...
    Genre:
      type: string
      enum:
        - BIOGRAPHY
        - HISTORY
        - SCIENCE
        - CRIME
        - POETRY
        - MYSTERY
        - SCIENCE_FICTION
        - FANTASY
        - ROMANCE
        - HORROR
        - DRAMA
      description: An enum with the possible book genres
      example: MYSTERY
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer Token authentication using JWT